사이드변경에 따른 매핑 정보 처리 수정 - 사이드사용여부 Y, 선택그룹 변경시 이전 선택그룹에 있는 상품코드매핑 데이터 삭제

+++ src/main/java/kr/co/solbipos/base/prod/prod/service/impl/ProdServiceImpl.java
@@ -503,25 +503,27 @@ public class ProdServiceImpl implements ProdService {
                 prodMapper.insertSdselProdToStore(prodVO);
             }
 
-            // 매핑스트링 TMP 테이블
-            ArtiseeProdMappingVO artiseeProdMappingVO = new ArtiseeProdMappingVO();
-            artiseeProdMappingVO.setHqOfficeCd(sessionInfoVO.getHqOfficeCd());
-            artiseeProdMappingVO.setSessionId(sessionInfoVO.getSessionId());
-            artiseeProdMappingVO.setUserId(sessionInfoVO.getUserId());
+            if(sessionInfoVO.getHqOfficeCd().equals("A0001")) {
+                // 매핑스트링 TMP 테이블
+                ArtiseeProdMappingVO artiseeProdMappingVO = new ArtiseeProdMappingVO();
+                artiseeProdMappingVO.setHqOfficeCd(sessionInfoVO.getHqOfficeCd());
+                artiseeProdMappingVO.setSessionId(sessionInfoVO.getSessionId());
+                artiseeProdMappingVO.setUserId(sessionInfoVO.getUserId());
 
-            // TMP테이블 삭제
-            artiseeProdMappingMapper.deleteMappingTmp01(artiseeProdMappingVO);
-            artiseeProdMappingMapper.deleteMappingTmp02(artiseeProdMappingVO);
+                // TMP테이블 삭제
+                artiseeProdMappingMapper.deleteMappingTmp01(artiseeProdMappingVO);
+                artiseeProdMappingMapper.deleteMappingTmp02(artiseeProdMappingVO);
 
-            // 맵핑스트링 저장
-            artiseeProdMappingMapper.insertMappingString(artiseeProdMappingVO);
+                // 맵핑스트링 저장
+                artiseeProdMappingMapper.insertMappingString(artiseeProdMappingVO);
 
-            // ERP 맵핑상품코드 삭제
-            SideMenuManageVO sideMenuManageVO = new SideMenuManageVO();
-            sideMenuManageVO.setHqOfficeCd(sessionInfoVO.getHqOfficeCd());
-            sideMenuManageVO.setSessionId(sessionInfoVO.getSessionId());
-            sideMenuManageVO.setUserId(sessionInfoVO.getUserId());
-            sideMenuMapper.getErpProdMappingDelete(sideMenuManageVO);
+                // ERP 맵핑상품코드 삭제
+                SideMenuManageVO sideMenuManageVO = new SideMenuManageVO();
+                sideMenuManageVO.setHqOfficeCd(sessionInfoVO.getHqOfficeCd());
+                sideMenuManageVO.setSessionId(sessionInfoVO.getSessionId());
+                sideMenuManageVO.setUserId(sessionInfoVO.getUserId());
+                sideMenuMapper.getErpProdMappingDelete(sideMenuManageVO);
+            }
         }
 
         // 상품 판매가 저장
